Move TransactionSet to the transactions module
