+++ sandbox/src/timestamping.rs
@@ -39,7 +39,9 @@ impl Transaction for TimestampTx {
         self.verify_signature(self.pub_key())
     }
 
-    fn execute(&self, _: &mut Fork) {}
+    fn execute(&self, _: &mut Fork) -> ExecutionResult {
+        Ok(())
+    }
 }
